fix response

+++ src/controllers/drones/updateDroneController.js
@@ -4,6 +4,12 @@ import path, { dirname, join } from 'path'
 import { DroneModel } from "../../models/Drones.js"
 
 const updateDroneController = async (req, res) => {
+    let response = {
+        code: 200,
+        message: "",
+        data: {},
+    }
+
     try {
         const { id } = req.params
         const { name, serialNumber, expiredDate } = req.body
@@ -11,17 +17,35 @@ const updateDroneController = async (req, res) => {
         const specifications = JSON.parse(req.body.specifications)
         const transponder = JSON.parse(req.body.transponder)
 
-        const dronePicture = req.files['dronePicture'][0]
-        const emergencyProcedure = req.files['emergencyProcedure'][0]
-        const insuranceDocument = req.files['insuranceDocument'][0]
-        const listOfEquipment = req.files['listOfEquipment'][0]
-        const droneCertificate = req.files['droneCertificate'][0]
+        const lastDrone = await DroneModel.findById({
+            _id: id,
+            auth: {
+                userId: req.userId,
+                role: req.role
+            }
+        })
+
+        let dronePicture = lastDrone.document.dronePicture
+        if (req.files['dronePicture']) dronePicture = req.files['dronePicture'][0]
+
+        let emergencyProcedure = lastDrone.document.emergencyProcedure
+        if (req.files['emergencyProcedure']) emergencyProcedure = req.files['emergencyProcedure'][0]
+
+        let insuranceDocument = lastDrone.document.insuranceDocument
+        if (req.files['insuranceDocument']) insuranceDocument = req.files['insuranceDocument'][0]
+
+        let listOfEquipment = lastDrone.document.listOfEquipment
+        if (req.files['listOfEquipment']) listOfEquipment = req.files['listOfEquipment'][0]
+
+        let droneCertificate = lastDrone.document.droneCertificate
+        if (req.files['droneCertificate']) droneCertificate = req.files['droneCertificate'][0]
+
         const newDocument = {
-            dronePicture: dronePicture.path.replace(/\\/g, '/'),
-            emergencyProcedure: emergencyProcedure.path.replace(/\\/g, '/'),
-            insuranceDocument: insuranceDocument.path.replace(/\\/g, '/'),
-            listOfEquipment: listOfEquipment.path.replace(/\\/g, '/'),
-            droneCertificate: droneCertificate.path.replace(/\\/g, '/'),
+            dronePicture: dronePicture === lastDrone.document.dronePicture ? dronePicture : dronePicture.path.replace(/\\/g, '/'),
+            emergencyProcedure: emergencyProcedure === lastDrone.document.emergencyProcedure ? emergencyProcedure : emergencyProcedure.path.replace(/\\/g, '/'),
+            insuranceDocument: insuranceDocument === lastDrone.document.insuranceDocument ? insuranceDocument : insuranceDocument.path.replace(/\\/g, '/'),
+            listOfEquipment: listOfEquipment === lastDrone.document.listOfEquipment ? listOfEquipment : listOfEquipment.path.replace(/\\/g, '/'),
+            droneCertificate: droneCertificate === lastDrone.document.droneCertificate ? droneCertificate : droneCertificate.path.replace(/\\/g, '/'),
         }
 
         const updatedData = {
@@ -34,7 +58,7 @@ const updateDroneController = async (req, res) => {
             updatedAt: new Date()
         }
 
-        const lastDrone = await DroneModel.findOneAndUpdate(
+        const newDrone = await DroneModel.findOneAndUpdate(
             {
                 _id: id,
                 auth: {
@@ -43,24 +67,44 @@ const updateDroneController = async (req, res) => {
                 }
             },
             updatedData,
+            { new: true }
         )
+
+        if (!newDrone) {
+            response.code = 404
+            response.message = "Drone not found"
+            response.data = {}
+            return res.status(404).json(response)
+        }
 
         const __filename = fileURLToPath(import.meta.url)
         const __dirname = dirname(__filename)
         const assetsDirectory = join(__dirname, '../../../')
-        fs.unlinkSync(join(assetsDirectory, lastDrone.document.dronePicture))
-        fs.unlinkSync(join(assetsDirectory, lastDrone.document.emergencyProcedure))
-        fs.unlinkSync(join(assetsDirectory, lastDrone.document.insuranceDocument))
-        fs.unlinkSync(join(assetsDirectory, lastDrone.document.listOfEquipment))
-        fs.unlinkSync(join(assetsDirectory, lastDrone.document.droneCertificate))
 
-        if (!lastDrone) {
-            return res.status(404).json({ error: 'Drone not found' })
-        }
+        if (dronePicture !== lastDrone.document.dronePicture)
+            fs.unlinkSync(join(assetsDirectory, lastDrone.document.dronePicture))
 
-        return res.status(200).json({ message: "Drone updated succesfull", lastDrone })
-    } catch (error) {
-        res.status(500).json({ error: "Internal server error", detail: error.message })
+        if (emergencyProcedure !== lastDrone.document.emergencyProcedure)
+            fs.unlinkSync(join(assetsDirectory, lastDrone.document.emergencyProcedure))
+
+        if (insuranceDocument !== lastDrone.document.insuranceDocument)
+            fs.unlinkSync(join(assetsDirectory, lastDrone.document.insuranceDocument))
+
+        if (listOfEquipment !== lastDrone.document.listOfEquipment)
+            fs.unlinkSync(join(assetsDirectory, lastDrone.document.listOfEquipment))
+
+        if (droneCertificate !== lastDrone.document.droneCertificate)
+            fs.unlinkSync(join(assetsDirectory, lastDrone.document.droneCertificate))
+
+        response.code = 200
+        response.message = "Drone data successfully update"
+        response.data = { drone: newDrone }
+        return res.status(200).json(response)
+    } catch (e) {
+        response.code = 500
+        response.message = e.message
+        response.data = {}
+        return res.status(500).json(response)
     }
 }
